pyproject.toml
[project]
name = "thinkdepthai_deep_research"
version = "0.1.0"
description = "ThinkDepth.ai Deep Research"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.5.4",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain_community>=0.3.27",
    "langchain_tavily>=0.2.7",
    "langchain_mcp_adapters>=0.1.9",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
    "tavily-python>=0.5.0",
    "duckdb>=1.4.2",
    "pytz>=2025.2",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deep_research"]

[tool.setuptools.package-dir]
"deep_research" = "src"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
